This is how it doesn't work.
In Dealer class, the way we've had it all along, the board adds the
player.
That seems to be why, if board has a BoxLayout or a GridLayout, and adds
the columns, and then the player, that's why it's shoved around weirdly.

If you comment out board.add(player); in the Dealer class, you can see
the columns laid out much more normally.